Clean up main.dart + add snake state notifier
