+++ pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
+  flutter_riverpod: ^2.3.2
 
 dev_dependencies:
   flutter_test:
